requirements.txt
streamlit==1.30.0
langchain-community==0.3.27
openai>=2.2.0
tiktoken
chromadb
python-docx
docx2txt>=0.8
scikit-learn
